+++ docker/docker-compose.yml
@@ -21,11 +21,4 @@ services:
     command:
       - start-dev
       - --import-realm
-  mockserver:
-    image: mockserver/mockserver:5.15.0
-    ports:
-      - "1080:1080"
-    environment:
-      MOCKSERVER_INITIALIZATION_JSON_PATH: /config/init_mockserver.json
-    volumes:
-      - ./mock-server/init_mockserver.json:/config/init_mockserver.json
+
