+++ src/main/java/com/forero/parking/application/port/EmailServerPort.java
@@ -1,6 +1,6 @@
 package com.forero.parking.application.port;
 
-import com.forero.parking.domain.model.Email;
+import com.forero.parking.infrastructure.adapter.gateways.Email;
 
 public interface EmailServerPort {
     String sendEmail(Email email);
